+++ src/gui/propspanel/GeneralML.ts
@@ -115,6 +115,7 @@ let genHTML = `
         <div>
                 <button id="swAv" title="use as avatar"><span class="material-icons-outlined">wc</span></button>
                 <button id="swGnd" title="use as ground"><span class="material-icons-outlined">terrain</span></button>
+                <button id="addCC" title="add a character controller to this ">CC</button>
         </div>